+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `--namespace` flag from `generate` command.
 - Remove defaulting to "giantswarm" from `--name` flag in `generate` command.
 
+### Fixed
+
+- Do not render `status:` in `generate` command.
+
 ## [0.2.0] - 2021-05-19
 
 ### Added